LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-active-unreached-target-expected.txt

1 2

FAIL target whose target position is not yet reached only get selected when its less than half a scroll port away. assert_equals: marker 1 should be unselected. expected "rgb(255, 0, 0)" but got ""
